requirements.txt
fastapi
uvicorn[standard]
python-dotenv
cohere
 